feat(mongodb): re-organizing the stepper Found out that the wizard is a bit weird and hard to change, so I split this in 4 different step components. Also providing a .env.example config file, so that we move the DB connection to .env
