+++ .env.example
@@ -1,0 +1,5 @@
+SINGLESTORE_HOST=
+SINGLESTORE_USER=
+SINGLESTORE_PASS=
+SINGLESTORE_DB=
+SINGLESTORE_DRIVER=mysql
