+++ .gitignore
@@ -26,7 +26,7 @@ yarn-error.log*
 .pnpm-debug.log*
 
 # local env files
-.env
+.env.local
 .env*.local
 
 # vercel
