+++ database/config/config.js
@@ -1,16 +1,7 @@
 export const dbDetails = {
-    HOST: "svc-46f38d9c-f13e-44c8-b3ea-9c0cb49adedf-dml.gcp-netherlands-1.svc.singlestore.com",
-    USER: "admin",
-    PASSWORD: "Beast*8640",
-    DB: "HackSingle",
-    dialect: "mysql",
+    HOST: process.env.SINGLESTORE_HOST,
+    USER: process.env.SINGLESTORE_USER,
+    PASSWORD: process.env.SINGLESTORE_PASS,
+    DB: process.env.SINGLESTORE_DB,
+    dialect: process.env.SINGLESTORE_DRIVER || "mysql",
 };
-
-// Mysql connection details
-// export const dbDetails = {
-//     HOST: "localhost",
-//     USER: "vincent",
-//     PASSWORD: "secret",
-//     DB: "NoviumLocal",
-//     dialect: "mysql",
-// };
